Rename `print-preprocess.js`
